fix: use neovim native solarized theming
This patch fixes a highlighting issue with the old Solarized Vim plugin
by replacing it with a more fully featured Neovim plugin.

There is still an issue with airline colors being gray when in tmux
(instead of the correct solarized background color), but that is
somewhat non-blocking and I will investigate further later.